Refactored pre_quant into base
