+++ awq/models/mpt.py
@@ -3,10 +3,10 @@ from .base import BaseAWQForCausalLM
 class MptAWQForCausalLM(BaseAWQForCausalLM):
     layer_type = "MPTBlock"
 
-    def get_model_layers(model):
+    def get_model_layers(self, model):
         return model.transformer.blocks
     
-    def get_layers_for_scaling(module, input_feat, module_kwargs):
+    def get_layers_for_scaling(self, module, input_feat, module_kwargs):
         layers = []
 
         # attention input
@@ -42,13 +42,13 @@ class MptAWQForCausalLM(BaseAWQForCausalLM):
 
         return layers
     
-    def get_act_for_scaling(module):
+    def get_act_for_scaling(self, module):
         return dict(
             scale_name="ffn.act",
             scale_layer=module.ffn.act,
             scale_shape=module.ffn.up_proj.out_features
         )
     
-    def move_embed(model, device):
+    def move_embed(self, model, device):
         model.transformer.wte = model.transformer.wte.to(device)
         model.transformer.emb_drop = model.transformer.emb_drop.to(device)
