+++ awq/utils/module.py
@@ -1,4 +1,7 @@
+import torch.nn as nn
 
+def get_named_linears(module):
+    return {name: m for name, m in module.named_modules() if isinstance(m, nn.Linear)}
 
 def get_op_by_name(module, op_name):
     # get the op by its name relative to the module
